requirements.txt
argparse==1.2.1
iso8601==0.1.4
prettytable==0.7.2
python-novaclient==2.13.0
requests==1.2.0
simplejson==3.1.3
wsgiref==0.1.2
